# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Answer API",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/commands/create-lesson": {
            "post": {
                "tags": [
                    "Lessons"
                ],
                "parameters": [
                    {
                        "parameter": "lesson_in_body",
                        "name": "Lesson",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Lesson"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Create a lesson"
                    }
                }
            }
        },
        "/lessons": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "get page of lessons",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search in lessons",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "course_id",
                        "in": "query",
                        "description": "filter the lessons by course",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "subject_id",
                        "in": "query",
                        "description": "filter by subject_id",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "after_date",
                        "in": "query",
                        "description": "filter by created at date after date provided",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "before_date",
                        "in": "query",
                        "description": "filter by created at date before date provided",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "filter by active or inactive lessons",
                        "required": false,
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Lesson"
                    }
                }
            }
        },
        "/commands/update-lesson-reference-id": {
            "patch": {
                "tags": [
                    "Lessons"
                ],
                "parameters": [
                    {
                        "parameter": "lesson_in_body",
                        "name": "Lesson",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/UpdateLessonReferenceId"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Update the lesson reference id"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Lesson": {
            "required": [],
            "properties": {
                "reference_id": {
                    "description": "Class LessonController",
                    "type": "string"
                },
                "lesson_type_id": {
                    "description": "Class LessonController",
                    "type": "string"
                },
                "sequence": {
                    "description": "Class LessonController",
                    "type": "integer"
                },
                "subject_id": {
                    "description": "Class LessonController",
                    "type": "string"
                },
                "unit_id": {
                    "description": "Class LessonController",
                    "type": "string"
                },
                "title": {
                    "description": "Class LessonController",
                    "type": "string"
                },
                "multi_line_title": {
                    "description": "Class LessonController",
                    "type": "string"
                },
                "description": {
                    "description": "Class LessonController",
                    "type": "string"
                },
                "start_page": {
                    "description": "Class LessonController",
                    "type": "integer"
                },
                "ending_page": {
                    "description": "Class LessonController",
                    "type": "integer"
                },
                "challenge_page": {
                    "description": "Class LessonController",
                    "type": "integer"
                },
                "practice_page": {
                    "description": "Class LessonController",
                    "type": "integer"
                },
                "drill_page": {
                    "description": "Class LessonController",
                    "type": "integer"
                },
                "status": {
                    "description": "Class LessonController",
                    "type": "boolean"
                }
            }
        },
        "UpdateLessonReferenceId": {
            "required": [],
            "properties": {
                "lesson_id": {
                    "description": "Class LessonController",
                    "type": "string"
                },
                "lesson_reference_id": {
                    "description": "Class LessonController",
                    "type": "string"
                }
            }
        }
    },
...
}